Working on cart

+++ lib/services/cart_services.dart
@@ -3,17 +3,13 @@ import 'package:ecommerce_app/models/cart_item.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final cartServiceProvider = Provider<CartServices>(
-  (ref){
-    return CartServices();
-  }
-);
+final cartServiceProvider = Provider<CartServices>((ref) {
+  return CartServices();
+});
 
 class CartServices {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
   final FirebaseAuth auth = FirebaseAuth.instance;
-  late final User? user = auth.currentUser;
-  late final String? userId = user?.uid;
 
   Stream<List<CartItem>> getCartItems(String userId) {
     return _firestore
@@ -26,19 +22,44 @@ class CartServices {
   }
 
   Future<void> addToCart(CartItem item) async {
-    await _firestore
+    final user = auth.currentUser;
+    if (user != null) {
+      await _firestore
+          .collection('users')
+          .doc(user.uid)
+          .collection('cart')
+          .add(item.toMap());
+    }
+  }
+
+   Future<void> updateCartItem(String userId, CartItem item) async {
+    final cartItemDoc = await _firestore
         .collection('users')
         .doc(userId)
         .collection('cart')
-        .add(item.toMap());
+        .where('productId', isEqualTo: item.productId)
+        .limit(1)
+        .get();
+
+    if (cartItemDoc.docs.isNotEmpty) {
+      await _firestore
+          .collection('users')
+          .doc(userId)
+          .collection('cart')
+          .doc(cartItemDoc.docs.first.id)
+          .update(item.toMap());
+    }
   }
 
   Future<void> removeFromCart(String itemId) async {
-    await _firestore
-        .collection('users')
-        .doc(userId)
-        .collection('cart')
-        .doc(itemId)
-        .delete();
+    final user = auth.currentUser;
+    if (user != null) {
+      await _firestore
+          .collection('users')
+          .doc(user.uid)
+          .collection('cart')
+          .doc(itemId)
+          .delete();
+    }
   }
 }
